+++ SpringBootDemo/src/main/java/com/aikang/mapper/PlanExcItemMapper.java
@@ -11,9 +11,9 @@ public interface PlanExcItemMapper {
 
 	List<PlanExcItem> getAllPlanExcItems(@Param("company") String company);
 	
-	List<PlanExcItem> getPlanExcItemsByHid(@Param("hid") Integer hid, @Param("company") String company);
+	List<PlanExcItem> getPlanExcItemsByHid(@Param("hid") Long hid, @Param("company") String company);
 	
-	int deleteExcItemByHid(@Param("hid") Integer hid, @Param("company") String company);
+	int deleteExcItemByHid(@Param("hid") Long hid, @Param("company") String company);
     
     int addExcItemByList(@Param("record") PlanExcItem[] record, @Param("company") String company);
 }
